Add noirVersion to Noir circuits (#71)
This PR updates all of our current Noir circuits (both in
`circuit_database/` and in `circuit_tutorials/`) to include the new
field from the Sindri manifest: `noirVersion`. This field allows users
to specify whether we should apply the Noir prover version [0.17.0,
0.18.0, 0.19.4, 0.22.0, 0.23.0]. This is the first of two PR's that
patches our existing circuits, which have been assuming the 0.17.0
prover version. The next PR introduces more circuits of other versions -
see #70

## Reviewer Testing:
- [x] For each compressed circuit in `circuit_database/noir`, submit to
the frontend (stage server) and ensure compilation is successful:
  - [ ] not_equal.tar.gz
  - [ ] not_equal.zip
  - [x] decision_tree.tar.gz
  - [x] decision_tree.zip
  - [x] billion_zk_vote.tar.gz
  - [x] billion_zk_vote.zip
- [x] Ensure the `sindri.json` file at
`circuit_tutorials/noir/pagerank/circuits` has a field for the
`noirVersion` and that it is set to `0.17.0`