Patch to Rust Quickstart (#80)
This PR restores the rust quickstart functionality.

## Testing Instructions
You can run the rust quickstart the same as previous instructions:
```
cd reference_code/quickstart rust
SINDRI_API_KEY=... cargo run
```
You should see the output of compile and proving (if you tried this with
the current main branch, you'd hit a runtime error during the first
iteration of the compile polling loop.)

## Bug Source
With the new public circuits feature, we introduced the "public" field
to the circuit detail response. That was problematic with the previous
quickstart because there was a struct that assumed the public field
would only be returned as a vec (i.e. from the proof detail response),
otherwise it would not be found. That allowed one function to poll both
the circuit and proof status without defining an enum to capture both
response schemas.

## Solution
Since this is the quickstart, defining a verbose enum that elaborates on
the entire response schemas seemed too in-depth. Rather, the poll
response is now captured in the `serde_json::Value` type. The value type
is used as a wild-card when you don't really know what JSON responses
will come back until runtime. Not the right choice for production or an
SDK - but I do think it's a good fit for a quickstart template.