Updating Axiom-v2 Tutorial Circuit (#83)
Similar to the Axiom-v3 PR from yesterday, this patch downgrades the
same rust package for a different circuit. Note that this circuit is
featured in one of our documentation walkthroughs, so I did not include
any `.tar.gz` or `.zip` files, because we do not include those for the
tutorial circuits.