Jolt tutorial (#100)
This PR adds the Jolt tutorial code to the public `sindri-resources`
repo. This tutorial will be referenced in the documentation for the Jolt
"how-to" guide. The documentation PR will follow approval of this PR
(Note that links that reference that "how-to" guide are included but
will not work right now. Both PRs will be merged at the same time, but
this PR will need to be reviewed first).

To test the circuit, navigate to
sindri-resources/circuit_tutorials/jolt/sha3_chain
Change sample.env to .env and enter your Sindri API key.
Run:

cargo run --release

---------

Co-authored-by: Graham Ullrich <graham@flyingcracker.com>